ResponseHandler에서 throw 대신 statusCode로 처리하게 변경. 근데 tester는 동작이 안댐
