+++ src/main.cpp
@@ -3,7 +3,7 @@
 #include "Logger.hpp"
 #include "Server.hpp"
 #include "ConfigException.hpp"
-#include "HTTPException.hpp"
+#include "HttpException.hpp"
 #include "SysException.hpp"
 
 static std::string makeConfigPath(int argc, char* argv[], bool& isOption);
